refactor: split `users/models.py` into individual files (#421)
* refactor: move models to model folder

* refactor: update imports in tests

* fix: revert accidental change

* chore: run black

* fix: fix issues reported by flake8

* fix: remove unused import

* Revert "refactor: move models to model folder"

This reverts commit 0bf760ad07fc6d5fcbe33af5ed5c90d436fff379.

* fix: revert changes to tests

* refactor: split models into individual files

* fix: revert newline

* fix: set __all__ in init.py

* chore: run black formatter
